Use signals emitted by Spyder to update plugin
Specifically, we use signals to update the python path and the default wdir.
The default wdir is now set to the project directory if a project is open,
or os.getcwd() otherwise.

As a consequence, the plugin does not need a reference to the main Spyder
window, so the class UnitTestWidgetInSpyder is no longer necessary.