Fix status_label if no results to show
status_label update is triggered by assignment of test results in
process_finished. In case of no results, process_finished is called
with an empty list of testresults. This empty list must also be
assigned in order to trigger the status_label update, otherwise the
status remains "Running tests ..." after finishing. So we have to
differentiate between None and [] in the if clause.