Fix abbreviator if name has parameters with dots
The name of a test is the dotted test function name possibly followed
by a list of parameters in brackets. These parameters can contain dots
as for example in test_unittestwidget_process_finished_updates_status_label
from test_unittestgui.py where the label of the second parameter ends
with a dot, which led to the wrong abbreviation 's.w.t.test_u.t.]'. This
is fixed by not taking into account the parameter list when abbreviating.