Fix pytest test filename path resolution
Pytest test filenames are relative to the rootdir. This rootdir is not
necessarily the directory from which discovery starts, i.e. wdir from
spyder-unittest's config. For details on how pytest determins the rootdir see
https://docs.pytest.org/en/latest/customize.html#initialization-determining-rootdir-and-inifile.
Filenames used to be resolved against wdir. If wdir is not the rootdir (as
described in the above link) this resulted in invalid resolved filenames.
This commit gets the rootdir in the pytest_report_header hook which is
called after all configuration is done and before any test collection
starts. This rootdir is then used for filename path resolution.