Make pytest TestResult message clearer
If an error occurs in the setup phase of a test, then the test itself
(call phase) will not be executed and the status remains '---'. If an
error occurs in the teardown phase after a passed test, then the status
will be 'passed'. In both cases, the test category will be FAIL (red).
In order to make this more easily comprehensible we prepend 'ERROR at
setup: ' or 'ERROR at teardown: ' to the message in these cases.