Supply better message for expected to fail tests
if no reason was given in the test. Pytest gives no message whatsoever
in this case (as opposed to skipped test without reason message). As
we color xfailed tests green and xpassed tests red it is necessary to
give some explanatory message as to why the test has the status xfailed
or xpassed to avoid confusion about the color.