Use test names relative to wdir, not rootdir
Before the introduction of PyTestWorker, test names were taken from the
output of `pytest -v` and those names are relative to wdir. After, the
plugin displayed node ids which are relative to the pytest root dir.
This adds unnecessary information if the root dir is a parent of wdir,
so this commit reverts back to the earlier behaviour.