CI tests: Sleep in places to prevent QThread error
GitHub CI tests show the error "QThread: Destroyed while thread is
still running" when exiting pytest. This error surfaced in Spyder
commit f934456, part of PR spyder-ide/spyder#24269 (Get user
environment variables asynchronously). Possibly the thread getting
the environment variables is still running when pytest finishes,
but this does not explain why main Spyder does not show this error.

The workaround in this commit is to sleep for five seconds at the
end of every test using main_window and also before creating,
opening and closing projects.